Add files via upload
Yepkit control module script.

Note that you do need ykush's software and executable to be able to run this script. In addition a valid path to the executable needs to be present.
The executable needs the .dll file, the .exe, and .bat file in order to run. 

Power cycling functions can be edited in terms of time requirements at a later date. This script is fully functional and has undergone numerous validation and verification tests.